refactor: modularize cli_layout.py
