+++ docs/DECISIONS.md
@@ -757,3 +757,7 @@
   - Problem: The bias flag still risked collapsing small VSPreview hints, and screenshot confirmation adds interactive friction with no remaining consumers.
   - Decision: Dropped both fields from the config schema/template/docs, simplified confirmation to auto-approve without screenshot prompts, removed bias application in the alignment runner, and updated tests to reflect the streamlined flow.
   - Verification: `.venv/bin/pyright --warnings`; `.venv/bin/ruff check`; `PYTEST_DISABLE_PLUGIN_AUTOLOAD=1 .venv/bin/pytest -q`
+- *2025-12-10:* fix(cli layout): restored CLI output fidelity after refactor.
+  - Problem: CLI runs exited early with a `NameError` (`report_override` typo in `frame_compare.run_cli`) and rendered layout text showed literal `[[...]]` markers (style span parser looked for `]]]` instead of the actual `]]` delimiters), breaking runner CLI output tests.
+  - Decision: Corrected `run_cli` to pass `report_enable_override` through to `RunRequest` and fixed `_apply_style_spans` in `src/frame_compare/layout/renderer.py` to consume `[[role]]...[[/]]` spans using the proper `]]` closing token so Rich markup is applied (and stripped when `--no-color`).
+  - Verification: `.venv/bin/pytest -q tests/runner/test_cli_entry.py::test_cli_applies_overrides_and_naming`; `.venv/bin/pytest -q tests/runner/test_audio_alignment_cli.py::test_audio_alignment_block_and_json`; `.venv/bin/pytest -q tests/cli/test_layout.py`
