+++ workspace/comparison_videos/probe/681cf32e3434c4f6fd37d8e2a6759f01b85ac3b5.json
@@ -1,0 +1,1 @@
+{"applied_fps": null, "cache_key": "681cf32e3434c4f6fd37d8e2a6759f01b85ac3b5", "cached_at": "2025-12-10T14:01:32.828013+00:00", "effective_fps": [24000, 1001], "fps_override": null, "metadata_digest": "88737c12bd97fcd25b8c3f9cfa8f9c8af2c8ae62", "schema_version": 1, "source_fps": [24000, 1001], "source_frame_props": {}, "source_height": 1080, "source_num_frames": 24000, "source_width": 1920, "tonemap_prop_keys": [], "trim_end": -12, "trim_start": 0}
