+++ workspace/comparison_videos/probe/f680c6b5ac9933f4441590af7d0877d113058778.json
@@ -1,0 +1,1 @@
+{"applied_fps": [24000, 1001], "cache_key": "f680c6b5ac9933f4441590af7d0877d113058778", "cached_at": "2025-12-10T14:01:32.828756+00:00", "effective_fps": [24000, 1001], "fps_override": null, "metadata_digest": "dbc783f4f141c87558c3a2066b24ce83cc621711", "schema_version": 1, "source_fps": [24000, 1001], "source_frame_props": {}, "source_height": 1080, "source_num_frames": 24000, "source_width": 1920, "tonemap_prop_keys": [], "trim_end": null, "trim_start": 5}
