+++ workspace/config/config.toml
@@ -1,0 +1,1 @@
+config = true
